refactor(#652): split workers/edge/app.ts by trust domain
app.ts was 433 lines mixing five responsibilities (Env definition,
path-classification predicates, the forwarding primitive, the anonymous
turnstile/rate-limit/budget flow, session migration, the image proxy,
and Hono assembly). Design: docs/iterations/iter6/design-C2-app-ts-split.md
(Option B, split by trust domain, keep the Hono assembly as the single
place route-registration order is visible).

New modules, all pure moves (no behavior change):
- env.ts: Env, WorkerExecutionContext
- routing-policy.ts: PUBLIC_V1/ANON_V1/AUTH_RATE_LIMITED_* tables +
  isPublicV1/isAnonymousV1/isAuthRateLimited/normalizeV1Path/decodedForRouting
  (the #479/#464 percent-decode comment travels with it)
- forward.ts: forwardV1/forwardPublicCatalog/publicCatalogHeaders/
  catalogOutbound/authenticatedForward
- anonymous-flow.ts: handleAnonymousV1/anonymousForward/guardBudget/
  withAnonymousCookie (the #447 Turnstile-ordering comment travels with it)
- session-migrate.ts: SESSION_MIGRATE_PATH/handleSessionMigrate + the
  full #507 decision-reversal comment
- image-proxy.ts: handleImageProxy
- responses.ts: UNAUTHORIZED_BODY/NOT_FOUND_BODY/unauthorized/
  logInvalidCredential, plus rateLimitedResponse (placed here rather than
  in anonymous-flow.ts as the design doc's list suggests: forward.ts's
  authenticatedForward also needs it, and anonymous-flow.ts already
  imports forwardV1 from forward.ts, so keeping rateLimitedResponse in
  anonymous-flow.ts would create a forward.ts <-> anonymous-flow.ts
  import cycle. responses.ts sits below both with no behavior change.)
- app.ts: now only registerWorkerRoutes/createWorkerApp (94 lines),
  re-exporting Env/catalogOutbound/isAuthRateLimited so every existing
  `from "./app.ts"` test import keeps working untouched.

Existing independent modules (auth.ts, tiles.ts, turnstile.ts,
rateLimiter.ts, costBreaker.ts, edgeGuard.ts, containerEnv.ts, entry.ts)
are untouched. The V1_TRUST_CHAIN pipeline + Admitted branded-type lock
from the design's v2 addendum is out of scope for this card (#652 asked
for migration + app.ts split only) and is left for a follow-up.

Also fixes apps/web/tests/integration/session-migration-on-login.test.ts,
which pinned SESSION_MIGRATE_PATH against app.ts's source text; updated
to read session-migrate.ts, where the literal now lives.

Verified: pnpm run test:worker (258/258 pass, unchanged before/after),
apps/web's session-migration-on-login.test.ts (6/6 pass via
vitest.integration.config.ts), packages/contract's anon-limits.test.ts
(11/11 pass), make check (Python lint + typecheck + 1724 unit tests
green; test-integration fails locally only on the same pre-existing
Atlas CLI version mismatch noted in the prior commit, unrelated to this
split), actionlint (clean), and the config-read-sets meta-check (OK).
New files are 21-94 lines each, well within the 300-line/file budget;
oxlint's pre-existing unused-var finding in authConfig.test.ts predates
this migration and is out of scope (workers/edge has no oxlint gate yet,
tracked separately as the L1 card).
