+++ workers/edge/env.ts
@@ -1,0 +1,21 @@
+/// <reference types="@cloudflare/workers-types" />
+import type { GuardNamespace } from "./guardStore.ts";
+import type { TileBucket } from "./tiles.ts";
+
+export interface Env {
+  CATALOG: { fetch: (req: Request) => Promise<Response> };
+  USERS: { fetch: (req: Request) => Promise<Response> };
+  CONTAINER: DurableObjectNamespace;
+  EDGE_GUARD: GuardNamespace;
+  MAP_TILES?: TileBucket;
+  SUPABASE_URL: string;
+  SUPABASE_ANON_KEY: string;
+  SUPABASE_SERVICE_ROLE_KEY: string;
+  /** Cloudflare Turnstile secret (Worker secret binding — never process.env). */
+  TURNSTILE_SECRET: string;
+  ANON_ACCESS_ENABLED?: string;
+  ANON_ID_SECRET?: string;
+  [key: string]: unknown;
+}
+
+export type WorkerExecutionContext = Pick<ExecutionContext, "waitUntil" | "passThroughOnException">;
