+++ workers/edge/image-proxy.ts
@@ -1,0 +1,26 @@
+import type { WorkerExecutionContext } from "./env.ts";
+
+/** Image proxy + cache for image.anitabi.cn (unchanged behaviour, ported from entry.js). */
+export async function handleImageProxy(request: Request, ctx: WorkerExecutionContext): Promise<Response> {
+  const imagePath = new URL(request.url).pathname.slice(5);
+  if (!imagePath || imagePath.includes("..")) return new Response("Bad request", { status: 400 });
+  const cacheKey = new Request(request.url, request);
+  const cached = await caches.default.match(cacheKey);
+  if (cached) return cached;
+  const upstream = await fetch(`https://image.anitabi.cn/${imagePath}`, {
+    headers: { "User-Agent": "Animichi/1.0" },
+  });
+  if (!upstream.ok) {
+    return new Response(upstream.body, {
+      status: upstream.status,
+      headers: { "Content-Type": upstream.headers.get("Content-Type") ?? "image/jpeg" },
+    });
+  }
+  const headers = new Headers(upstream.headers);
+  headers.set("Cache-Control", "public, max-age=604800, s-maxage=2592000");
+  headers.set("Access-Control-Allow-Origin", "*");
+  headers.delete("Set-Cookie");
+  const response = new Response(upstream.body, { status: 200, headers });
+  ctx.waitUntil(caches.default.put(cacheKey, response.clone()));
+  return response;
+}
